+++ docs/src/examples/QasGalleryCard/Box.vue
@@ -17,7 +17,11 @@ export default {
   computed: {
     props () {
       return {
-        name: 'Meu titulo',
+        headerProps: {
+          labelProps: {
+            label: 'Meu titulo'
+          }
+        },
         url: 'https://picsum.photos/200/300.jpg'
       }
     }