+++ ui/src/components/input/QasInput.vue
@@ -20,7 +20,7 @@
 <script>
 import { getRequiredLabel, getPlaceholder } from '../../helpers'
 
-import { defineAsyncComponent } from 'vue'
+import { defineAsyncComponent, computed } from 'vue'
 
 const Masks = {
   CompanyDocument: 'company-document',
@@ -35,6 +35,21 @@ export default {
 
   components: {
     QasCopy: defineAsyncComponent(() => import('../copy/QasCopy.vue'))
+  },
+
+  provide () {
+    return {
+      /**
+       * @see QasBtn.vue - Injetando os valores padrões para o QasBtn.
+       */
+      btnPropsDefaults: computed(() => {
+        return {
+          size: 'md',
+          variant: 'tertiary',
+          ...(this.hasError && { color: 'negative' })
+        }
+      })
+    }
   },
 
   inheritAttrs: false,
@@ -112,7 +127,7 @@ export default {
 
   computed: {
     hasError () {
-      return this.inputReference.hasError
+      return this.inputReference?.hasError
     },
 
     masks () {