+++ ui/src/components/table-generator/QasTableGenerator.vue
@@ -64,6 +64,17 @@ export default {
     QasCheckbox
   },
 
+  provide () {
+    return {
+      /**
+       * @see QasBtn.vue - Injetando os valores padrões para o QasBtn.
+       */
+      btnPropsDefaults: {
+        size: 'md'
+      }
+    }
+  },
+
   props: {
     actionsMenuProps: {
       default: undefined,