+++ ui/src/components/uploader/private/PvUploaderGalleryCard.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <qas-gallery-card v-bind="defaultGalleryCardProps">
+    <qas-gallery-card v-if="useGalleryCard" v-bind="defaultGalleryCardProps">
       <template v-if="hasGenerator" #bottom>
         <div>
           <qas-grid-generator v-if="hasGridGenerator" v-bind="defaultGridGeneratorProps" />
@@ -9,6 +9,10 @@
         </div>
       </template>
     </qas-gallery-card>
+
+    <qas-box v-else v-bind="cardBoxProps">
+      <qas-header v-bind="cardHeaderProps" />
+    </qas-box>
 
     <qas-dialog v-model="showDialog" use-form use-validation-all-at-once v-bind="dialogProps" @validate="onValidate">
       <template #description>
@@ -23,6 +27,8 @@ import QasDialog from '../../dialog/QasDialog.vue'
 import QasFormGenerator from '../../form-generator/QasFormGenerator.vue'
 import QasGalleryCard from '../../gallery-card/QasGalleryCard.vue'
 import QasGridGenerator from '../../grid-generator/QasGridGenerator.vue'
+import QasBox from '../../box/QasBox.vue'
+import QasHeader from '../../header/QasHeader.vue'
 
 import downloadFile from '../../../helpers/download-file.js'
 
@@ -30,10 +36,18 @@ export default {
   name: 'PvUploaderGalleryCard',
 
   components: {
+    QasBox,
     QasDialog,
     QasFormGenerator,
     QasGalleryCard,
-    QasGridGenerator
+    QasGridGenerator,
+    QasHeader
+  },
+
+  inject: {
+    isBox: {
+      default: false
+    }
   },
 
   inheritAttrs: false,
@@ -94,6 +108,11 @@ export default {
 
     useObjectModel: {
       type: Boolean
+    },
+
+    useGalleryCard: {
+      type: Boolean,
+      default: true
     }
   },
 
@@ -145,7 +164,6 @@ export default {
 
         buttonProps: btnProps,
         errorMessage: error,
-        errorIcon: icon,
 
         ...actionsMenuProps
       } = this.normalizedCardGalleryProps
@@ -154,20 +172,15 @@ export default {
        * Quando hasError for "true", significa que é falha ao enviar o arquivo ao servidor (upload), e nestes casos:
        *
        * buttonProps: deve sempre ser possível excluir a imagem, por isso o disable do botão é "false".
-       * errorMessage: o label do erro deve ser "Falha ao carregar arquivo."
-       * errorIcon: o ícone do erro deve ser "sym_r_cancel".
        */
       const buttonProps = this.hasError ? { ...btnProps, disable: false } : btnProps
-      const errorMessage = this.hasError ? 'Falha ao carregar arquivo.' : error || this.fileType
-      const errorIcon = this.hasError ? 'sym_r_cancel' : icon
 
       return {
         disable: this.hasError,
 
         ...this.normalizedCardGalleryProps,
 
-        errorIcon,
-        errorMessage,
+        errorMessage: 'Falha ao carregar o arquivo.',
 
         imageProps: {
           ...imageProps,
@@ -180,58 +193,66 @@ export default {
           }
         },
 
-        actionsMenuProps: {
-          ...actionsMenuProps,
-
-          buttonProps: {
-            disable: false,
-            ...buttonProps
+        headerProps: {
+          labelProps: {
+            label: this.normalizedModelValue.name,
+            ...(this.hasError && { color: 'negative' })
           },
 
-          list: {
-            ...(
-              this.hasEditButton &&
-              {
-                edit: {
-                  label: 'Editar',
-                  icon: 'sym_r_edit',
+          actionsMenuProps: {
+            ...actionsMenuProps,
 
-                  // callback
-                  handler: () => {
-                    this.dialogValues = { ...this.currentModelValue }
-                    this.toggleDialog()
-                  }
-                }
-              }
-            ),
-
-            ...(
-              this.hasDownloadButton &&
-              {
-                download: {
-                  label: 'Baixar arquivo',
-                  icon: 'sym_r_download',
-
-                  // callback
-                  handler: () => downloadFile(this.normalizedModelValue)
-                }
-              }
-            ),
-
-            destroy: {
-              label: 'Excluir',
-              color: 'grey-10',
-              icon: 'sym_r_delete',
-
-              // callback
-              handler: () => this.$emit('remove')
+            buttonProps: {
+              disable: false,
+              ...buttonProps,
+              ...(this.hasError && { color: 'negative' })
             },
 
-            ...list
+            list: {
+              ...(
+                this.hasEditButton &&
+                {
+                  edit: {
+                    label: 'Editar',
+                    icon: 'sym_r_edit',
+
+                    // callback
+                    handler: () => {
+                      this.dialogValues = { ...this.currentModelValue }
+                      this.toggleDialog()
+                    }
+                  }
+                }
+              ),
+
+              ...(
+                this.hasDownloadButton &&
+                {
+                  download: {
+                    label: 'Baixar arquivo',
+                    icon: 'sym_r_download',
+
+                    // callback
+                    handler: () => downloadFile(this.normalizedModelValue)
+                  }
+                }
+              ),
+
+              destroy: {
+                label: 'Excluir',
+                color: 'grey-10',
+                icon: 'sym_r_delete',
+
+                // callback
+                handler: () => this.$emit('remove')
+              },
+
+              ...list
+            }
           }
         },
 
-        ...this.normalizedModelValue
+        url: this.normalizedModelValue?.url
       }
     },
 
@@ -255,10 +276,6 @@ export default {
 
     fileName () {
       return this.url ? this.url.split('/').pop() : ''
-    },
-
-    fileType () {
-      return this.fileName.split('.').pop()
     },
 
     formFields () {
@@ -324,12 +341,49 @@ export default {
            * Obs: esta URL fake é usada apenas para aparecer a imagem de "Falha ao enviar arquivo."
            * e não é adicionada ao model.
            */
-          : { url: this.url || 'error-on-upload-file', ...this.file }
+          : { ...this.file, url: this.file.url || this.url || 'error-on-upload-file' }
       )
     },
 
     url () {
-      return this.normalizedCardGalleryProps?.url
+      return this.defaultGalleryCardProps?.url
+    },
+
+    cardBoxProps () {
+      const isBordered = (this.isBox || this.hasError)
+
+      return {
+        outlined: isBordered,
+        unelevated: isBordered
+      }
+    },
+
+    cardHeaderProps () {
+      return {
+        actionsMenuProps: {
+          useLabel: false,
+          ...this.defaultGalleryCardProps.headerProps?.actionsMenuProps
+        },
+
+        labelProps: {
+          label: this.file.name || this.fileName,
+          typography: 'h5',
+
+          ...(this.hasError && { color: 'negative' })
+        },
+
+        ...(this.hasError && { description: 'Falha ao carregar o arquivo.' }),
+
+        spacing: 'none',
+        useEllipsis: true
+      }
+    },
+
+    /**
+     * Indica se há slot customizado para exibir o erro somente nos casos onde:
+     */
+    hasCustomErrorSlot () {
+      return !this.hasError && this.file.isUploaded
     }
   },