+++ CHANGELOG.md
@@ -14,10 +14,82 @@ Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de s
 Devemos adicionar o comentário `<!-- N/A -->` (Não adicionar), para que não precise adicionar um item do changelog ao lançar uma nova versão stable.
 Caso adicionado no escopo inicial, todos os conteúdos abaixo não serão adicionados. Caso adicionado na linha, será considerado apenas ela.
 
-## Não publicado
+## [3.19.0-beta.6] - 10-09-2025
+## BREAKING CHANGES
+- `QasGalleryCard`:
+  - removido propriedades `name` e `actionsMenuProps` em favor de usar a nova propriedade `headerProps`.
+  - removido slot `header`, `actions` e `name`.
+  - modificado slot `image-error` para `image-error-message` para o nome fazer mais sentido com o que de fato faz.
+- rever todos locais para passar a propriedade `galleryCardProps` no novo padrão do `QasGalleryCard`:
+  - QasGalleryCard.
+  - QasUploader.
+  - QasGallery.
+- Rever todos locais que usam `QasBtn` dentro de componentes de usam `QasInput` ou `QasTableGenerator` pois elas alteram a prop `size` default via provide `btnPropsDefaults`.
+- Muitas mudanças no `QasUploader`, verificar se nada quebra e remover validações manuais referentes a propriedades `accept` e `maxFileSize`.
+
 ### Adicionado
 - `QasActions`: adicionado possibilidade de passar props dos botões sem necessariamente precisar abrir slot. ([#1332](https://github.com/bildvitta/asteroid/issues/1332))
 - `QasInfiniteScroll`: adicionado função `fetchList` no expose. ([#1315](https://github.com/bildvitta/asteroid/issues/1315))
+- `QasUploader`:
+  - adicionado propriedades:
+    - useEmptyResultText
+    - useGalleryCard
+    - useHeader
+    - maxFileSize (ja existia no QUploader mas agora foi definida no componente).
+  - adicionado novo slot `bottom-list`.
+  - adicionado evento `rejected` (ja existia no QUploader mas agora foi definida no componente).
+  - adicionado `files` via parâmetros no método interno `addFiles`.
+  - adicionado validações nativas pom feedback `notify` para props `accept` e `maxFileSize`.
+- `QasBtn`: adicionado propriedade `size`. ([#1345](https://github.com/bildvitta/asteroid/issues/1345))
+- `ui/src/css/mixins/set-button.scss`: adicionado novo parâmetro `size`.
+- `QasGalleryCardProps`:
+  - adicionado propriedade `headerProps`.
+  - adicionado novo slot `image-error-container`.
+- `QasGallery`:
+  - adicionado nova propriedade `showMoreAlign`.
+  - adicionado efeito no hover da img do QasGalleryCard para feedback visual.
+- `QasHeader`: adicionada propriedade `useEllipsis`.
+- `QasStepper`:
+  - Adicionado prop `useVertical` para ter possibilidade de utilizá-lo na vertical.
+  - Adicionado prop `headerNav` (substituído o attrs) para controle de navegação entre as steps.
+
+### Corrigido
+- `QasUploader`:
+  - corrigido exibição do QasGalleryCard quando um upload falha.
+  - corrigido exibição no QasGalleryCard para arquivos que não carregam no QImg, ex: PDF, DOCX.
+- `QasGallery`: corrigido elemento "ver mais", que estava alocado em um elemento com gutter que gerava inconsistência de tamanho.
+
+### Modificado
+- `QasUploader`:
+  - modificado layout da área de drag and drop.
+  - modificado comportamento da propriedade `columns` que antes fazia um merge com o que era passado de fora do componente e o default interno, agora o default só existe caso não seja passado pelo lado de fora.
+- `ui/src/plugins/notify-error/NotifyError.js`: modificado ícone de erro para `sym_r_error`.
+`QasBtn`:
+  - modificado cursor `disabled` para `progress` quando o botão esta com estado `loading`.
+  - modificações para propriedades `color`, `size` e `variant` terem padrões defaults configuráveis via provide `btnPropsDefaults`.
+- Modificado `QTooltip` para `QasTooltip`:
+  - QasCopy.
+  - QasTransfer.
+- `QasFilters`: alterado size do botão dentro do menu para `sm`.
+- `QasInput`: alterado propriedades padrões do QasBtn via provide `btnPropsDefaults`:
+  - size.
+  - variant.
+  - color.
+- `QasTableGenerator`: alterado propriedade padrão `size` do QasBtn via provide `btnPropsDefaults`.
+- `QasGalleryCard`:
+  - modificado header para utilizar componente `QasHeader` através da propriedade `headerProps`.
+  - modificado default da prop `errorIcon` para `sym_r_error`.
+  - modificado default da prop `errorMessage` para `Falha ao carregar imagem.`.
+  - modificado slot `image-error` para `image-error-message` para o nome fazer mais sentido com o que de fato faz.
+  - modificado layout para outlined quando estiver dentro do QasDialog.
+- `QasGallery`: mudanças necessárias para se adaptar as mudanças do `QasGalleryCard`.
+- `QasHeader`: modificado logica para não mostrar elemento do header em casos de conteúdos vazio pela computada `hasHeaderContent`.
+- `QasStepper`: Modificado exibição da cor primária nas linhas de cada step.
+
+### Removido
+- `QasGalleryCard`:
+  - removido propriedades `name` e `actionsMenuProps` em favor de usar a nova propriedade `headerProps`.
+  - removido slot `header`, `actions` e `name`.
 
 ### Corrigido
 - `QasFormGenerator`: corrigido erro ao não passar a propriedade `columns` quando há subset. ([#1299](https://github.com/bildvitta/asteroid/issues/1299))
@@ -4259,3 +4331,4 @@ Adicionado suporte para Pinia/Vuex Seguindo os padrões da biblioteca `@bildvitt
 [3.19.0-beta.3]: https://github.com/bildvitta/asteroid/compare/v3.19.0-beta.3-alpha.3...v3.19.0-beta.3?expand=1
 [3.19.0-beta.4]: https://github.com/bildvitta/asteroid/compare/v3.19.0-beta.3...v3.19.0-beta.4?expand=1
 [3.19.0-beta.5]: https://github.com/bildvitta/asteroid/compare/v3.19.0-beta.5-alpha.0...v3.19.0-beta.5?expand=1
+[3.19.0-beta.6]: https://github.com/bildvitta/asteroid/compare/v3.19.0-beta.6-alpha.7...v3.19.0-beta.6?expand=1